QuantisedAtomic --> ConfigureAtomic
