+++ test/InitialConditions/quantised_atomic.jl
@@ -1,6 +1,6 @@
 using Test
 using NQCDynamics
-using NQCDynamics.InitialConditions.QuantisedAtomic
+using NQCDynamics.InitialConditions.ConfigureAtomic
 using LinearAlgebra: norm
 using Unitful
 using UnitfulAtomic
@@ -9,4 +9,4 @@ using UnitfulAtomic
 atoms = Atoms([:O])
 model = Free()
 sim = Simulation(atoms, model)
-surface = QuantisedAtomic.SurfaceParameters(sim.atoms.masses, [1], Matrix{Float64}(undef, 3, 0), 10.0, [0, 0, 1.0])
+surface = ConfigureAtomic.SurfaceParameters(sim.atoms.masses, [1], Matrix{Float64}(undef, 3, 0), 10.0, [0, 0, 1.0])
